added security to admin

+++ products/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect, reverse,  get_object_or_404 
+from django.shortcuts import render, redirect, reverse,  get_object_or_404
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
@@ -29,7 +29,7 @@ def all_products(request):
                 if direction == 'desc':
                     sortkey = f'-{sortkey}'
             products = products.order_by(sortkey)
-            
+      
         if 'category' in request.GET:
             categories = request.GET['category'].split(',')
             products = products.filter(category__name__in=categories)
@@ -78,20 +78,20 @@ def add_product(request):
     """
     if not request.user.is_superuser:
         messages.error(request, 'Sorry, only for store owners')
-        return redirect(reverse('home/'))
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
             form.save()
             product = form.save()
-            messages.success(request,'Successfully added product!')
+            messages.success(request, 'Successfully added product!')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
             messages.error(request, 'Failed to add product. Please ensure the form is valid.')
     else:
         form = ProductForm()
-        
+
     template = 'products/add_product.html'
     context = {
         'form': form,
@@ -104,6 +104,10 @@ def edit_product(request, product_id):
     """
     Edit a product in the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only for store owners')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.files, instance=product)
@@ -116,7 +120,7 @@ def edit_product(request, product_id):
     else:
         form = ProductForm(instance=product)
         messages.info(request, f'You are editing {product.name}')
-    
+
     template = 'products/edit_product.html'
     context = {
         'form': form,
@@ -126,9 +130,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 @login_required
-def delete_product(request, template, context):
+def delete_product(request, product_id):
+    """ 
+    Delete product from the store 
+    """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only for store owners')
+        return redirect(reverse('home'))
 
-    product = get_object_or_404(Product,  pk=product_id)
+    product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
     return redirect(reverse('products'))
